seed database with 80 numbers

+++ database/migrations/2023_12_20_192635_create_transaction_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('transaction', function (Blueprint $table) {
             $table->id();
-            $table->decimal('amount');
+            $table->float('amount', 8, 2);
             $table->boolean('is_paid');
             $table->json('selected_numbers');
             $table->foreignId('user_id')->constrained();
